+++ otx/cli/tools/build.py
@@ -42,14 +42,17 @@ def get_args():
         "--train-data-roots",
         help="Comma-separated paths to training data folders.",
     )
+    parser.add_argument("--train-ann-files", help="Comma-separated paths to train annotation files.")
     parser.add_argument(
         "--val-data-roots",
         help="Comma-separated paths to validation data folders.",
     )
+    parser.add_argument("--val-ann-files", help="Comma-separated paths to train annotation files.")
     parser.add_argument(
         "--test-data-roots",
         help="Comma-separated paths to test data folders.",
     )
+    parser.add_argument("--test-ann-files", help="Comma-separated paths to train annotation files.")
     parser.add_argument(
         "--unlabeled-data-roots",
         help="Comma-separated paths to unlabeled data folders",