+++ tests/unit/core/data/manager/test_dataset_manager.py
@@ -97,6 +97,17 @@ class TestOTXDatasetManager:
         data_format = DATA_ROOTS2FORMAT[data_root]
         assert DatasetManager.import_dataset(data_root, data_format=data_format) is not None
 
+    # TODO: Currently, direct annotation only supports COCO format
+    @e2e_pytest_unit
+    @pytest.mark.parametrize("data_root", [AVAILABLE_DATA_ROOTS[1]])
+    def test_import_dataset_with_direct_annotation(self, data_root):
+        data_format = DATA_ROOTS2FORMAT[data_root]
+        assert DatasetManager.import_dataset(data_root, data_format=data_format) is not None
+
+        ann_files = "tests/assets/car_tree_bug/annotations/instances_train_5_imgs.json"
+        train_dataset = DatasetManager.import_dataset(ann_files, data_format=data_format, subset="train")
+        assert train_dataset.get_subset("train").get_annotated_items() == 5
+
     @e2e_pytest_unit
     @pytest.mark.parametrize("task", AVAILABLE_TASKS)
     @pytest.mark.parametrize("subset", AVAILABLE_SUBSETS)